drop whitelist from capacity & mv to core.
