+++ test/utils/DeployDealSystem.sol
@@ -32,8 +32,8 @@ library DeployDealSystem {
     // ------------------ Constants ------------------
     uint256 public constant DEFAULT_EPOCH_DURATION = 1 days;
     uint256 public constant DEFAULT_FLT_PRICE = 10 * PRECISION; // 10 USD
-    uint256 public constant DEFAULT_MIN_DEPOSITED_Epochs = 2;
-    uint256 public constant DEFAULT_MIN_REMATCHING_Epochs = 2;
+    uint256 public constant DEFAULT_MIN_DEPOSITED_EPOCHS = 2;
+    uint256 public constant DEFAULT_MIN_REMATCHING_EPOCHS = 2;
     uint256 public constant DEFAULT_MIN_PROTOCOL_VERSION = 1;
     uint256 public constant DEFAULT_MAX_PROTOCOL_VERSION = 1;
     uint256 public constant DEFAULT_USD_COLLATERAL_PER_UNIT = 100 * PRECISION; // 100 USD
@@ -46,7 +46,7 @@ library DeployDealSystem {
     uint256 public constant DEFAULT_SLASHING_RATE = 100000; // 0.01 = 1% = 100000
     uint256 public constant DEFAULT_MIN_REQUIERD_PROOFS_PER_EPOCH = 3;
     uint256 public constant DEFAULT_MAX_PROOFS_PER_EPOCH = 5;
-    uint256 public constant DEFAULT_WITHDRAW_Epochs_AFTER_FAILED = 2;
+    uint256 public constant DEFAULT_WITHDRAW_EPOCHS_AFTER_FAILED = 2;
     uint256 public constant DEFAULT_MAX_FAILED_RATIO = 10;
     uint256 public constant DEFAULT_INIT_REWARD_POOL = 1000 ether;
     bool public constant DEFAULT_IS_WHITELIST_ENABLED = false;
@@ -70,11 +70,12 @@ library DeployDealSystem {
                     abi.encodeWithSelector(
                         Core.initialize.selector,
                         DEFAULT_EPOCH_DURATION,
-                        DEFAULT_MIN_DEPOSITED_Epochs,
-                        DEFAULT_MIN_REMATCHING_Epochs,
+                        DEFAULT_MIN_DEPOSITED_EPOCHS,
+                        DEFAULT_MIN_REMATCHING_EPOCHS,
                         DEFAULT_MIN_PROTOCOL_VERSION,
                         DEFAULT_MAX_PROTOCOL_VERSION,
-                        dealImpl
+                        dealImpl,
+                        DEFAULT_IS_WHITELIST_ENABLED
                     )
                 )
             )
@@ -105,9 +106,8 @@ library DeployDealSystem {
                         DEFAULT_SLASHING_RATE,
                         DEFAULT_MIN_REQUIERD_PROOFS_PER_EPOCH,
                         DEFAULT_MAX_PROOFS_PER_EPOCH,
-                        DEFAULT_WITHDRAW_Epochs_AFTER_FAILED,
+                        DEFAULT_WITHDRAW_EPOCHS_AFTER_FAILED,
                         DEFAULT_MAX_FAILED_RATIO,
-                        DEFAULT_IS_WHITELIST_ENABLED,
                         DEFAULT_INIT_GLOBAL_NONCE,
                         DEFAULT_DIFFICULTY_TARGET,
                         DEFAULT_INIT_REWARD_POOL,
